Namespace PgQueryMethods under Mobility::Sequel
